add solvation tests for charmm
