+++ mdpow/tests/testing_resources/molecules/benzene/benzene_charmm.itp
@@ -1,0 +1,109 @@
+; Created by cgenff_charmm2gmx.py
+
+[ moleculetype ]
+; Name			   nrexcl
+BNZ				 3
+
+[ atoms ]
+;	 nr		  type	resnr residue  atom   cgnr	   charge		mass  typeB    chargeB		massB
+; residue	 1 BNZ rtp BNZ q	qsum
+     1     CG2R61      1    BNZ     C1      1     -0.115     12.011   ;
+     2      HGR61      1    BNZ     H7      2      0.115      1.008   ;
+     3     CG2R61      1    BNZ     C2      3     -0.115     12.011   ;
+     4      HGR61      1    BNZ     H8      4      0.115      1.008   ;
+     5     CG2R61      1    BNZ     C3      5     -0.115     12.011   ;
+     6      HGR61      1    BNZ     H9      6      0.115      1.008   ;
+     7     CG2R61      1    BNZ     C4      7     -0.115     12.011   ;
+     8      HGR61      1    BNZ    H10      8      0.115      1.008   ;
+     9     CG2R61      1    BNZ     C5      9     -0.115     12.011   ;
+    10      HGR61      1    BNZ    H11     10      0.115      1.008   ;
+    11     CG2R61      1    BNZ     C6     11     -0.115     12.011   ;
+    12      HGR61      1    BNZ    H12     12      0.115      1.008   ;
+
+[ bonds ]
+;	ai	  aj funct			  c0			c1			  c2			c3
+    1     2	 1
+    1     3	 1
+    1    11	 1
+    3     4	 1
+    3     5	 1
+    5     6	 1
+    5     7	 1
+    7     9	 1
+    7     8	 1
+    9    10	 1
+    9    11	 1
+   11    12	 1
+
+[ pairs ]
+;	ai	  aj funct			  c0			c1			  c2			c3
+    1    10	 1
+    1     6	 1
+    1     7	 1
+    2     9	 1
+    2    12	 1
+    2     4	 1
+    2     5	 1
+    3     9	 1
+    3    12	 1
+    3     8	 1
+    4    11	 1
+    4     6	 1
+    4     7	 1
+    5    11	 1
+    5    10	 1
+    6     9	 1
+    6     8	 1
+    7    12	 1
+    8    10	 1
+    8    11	 1
+   10    12	 1
+
+[ angles ]
+;	ai	  aj	ak funct			c0			  c1			c2			  c3
+    2     1     3	5
+    2     1    11	5
+    3     1    11	5
+    1     3     4	5
+    1     3     5	5
+    4     3     5	5
+    3     5     6	5
+    3     5     7	5
+    6     5     7	5
+    9     7     5	5
+    9     7     8	5
+    5     7     8	5
+   10     9    11	5
+   10     9     7	5
+   11     9     7	5
+    9    11     1	5
+    9    11    12	5
+    1    11    12	5
+
+[ dihedrals ]
+;	ai	  aj	ak	  al funct			  c0			c1			  c2			c3			  c4			c5
+    2     1     3     4	 9
+    2     1     3     5	 9
+   11     1     3     4	 9
+   11     1     3     5	 9
+    2     1    11     9	 9
+    2     1    11    12	 9
+    3     1    11     9	 9
+    3     1    11    12	 9
+    1     3     5     6	 9
+    1     3     5     7	 9
+    4     3     5     6	 9
+    4     3     5     7	 9
+    3     5     7     9	 9
+    3     5     7     8	 9
+    6     5     7     9	 9
+    6     5     7     8	 9
+    5     7     9    10	 9
+    5     7     9    11	 9
+    8     7     9    10	 9
+    8     7     9    11	 9
+   10     9    11     1	 9
+   10     9    11    12	 9
+    7     9    11     1	 9
+    7     9    11    12	 9
+
